feat(twitter): ported to modern api library
The Twitter API is moving towards a v2 as we speak. The previous
library used is several years old and wouldn't be able to adapt.

The new library has Typescrypt type definitions built in, is self
contained with no dependencies, and has an active developer that
frequently fixes and updates the API as new features are released.